Make client optional
