+++ Cargo.toml
@@ -8,7 +8,7 @@ authors = ["Allen Bui <allenbui94@gmail.com>"]
 license = "MIT"
 
 [dependencies]
-reqwest = "0.9"
+reqwest = { version = "0.9", optional = true }
 http = "0.1"
 serde = "1"
 serde_json = "1.0.11"
@@ -17,3 +17,7 @@ log = "0.3"
 
 [dev-dependencies]
 kankyo = "~0.2"
+
+[features]
+default = ["reqwest-client"]
+reqwest-client = ["reqwest"]
